fix(kmm): Better handling of port conflicts
* feat(common): Implement `PortConflictException`

* feat(kmm): Implement `SocketIo`

Implement `SocketIo` to determine if an arbitrary port is available.

* feat(kmm): Add `SocketIo` to dependency injection

* feat(kmm): Throw custom exception upon port conflict

Upon calling `startServer()`, throw a `PortConflictException` if the desired port is already in use.

* style(kmm): Rename `SocketIoUtils` -> `MockSelectorManager`

Fix spotless issues

* refactor(kmm): Refactor `SocketIo` dependencies

Provider `SocketBinder` interface to `SocketIo` instead of `SelectorManager` to allow for mocking in unit tests.

* fix(kmm): Update `SocketIo` for special handling of port 0

Ktor treats port 0 as a flag to use a random, available port. As such, `SocketIo.isPortAvailable` should always return `true` for this port.