fix: Try synchronously starting the update server block so the existing job is always cancelled before triggering a state update
Beforehand it would be possible though unlikely that the update server block was not executed in time to be able to cancel the job for server updates before the update state block updated the view model's state (with an old 600 millisecond value). This could have been a cause for editor synchronisation issues. We can ensure this doesn't happen by making the cancel for the job always run before the view model state update, which means we'll always refresh the 600 millisecond debounce timer before it might conflict with an old value.